+++ backend/users/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from .models import CustomUser, UserInventory, Deposit, Withdrawal
+from .models import CustomUser, UserInventory
 
 @admin.register(CustomUser)
 class UserAdmin(admin.ModelAdmin):
@@ -8,13 +8,5 @@ class UserAdmin(admin.ModelAdmin):
 @admin.register(UserInventory)
 class UserInventoryAdmin(admin.ModelAdmin):
     list_display = ["user", "pack"]
-
-@admin.register(Deposit)
-class DepositAdmin(admin.ModelAdmin):
-    list_display = ["user", "wallet", "sum", "date"]
-
-@admin.register(Withdrawal)
-class WithdrawalAdmin(admin.ModelAdmin):
-    list_display = ["user", "sum", "date"]
